refactor: modularize network security
