+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -162,6 +162,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Brought down from 892 to 148 lines by modularizing handshake, routing, authentication, and logging utilities. Protocol file now orchestrates these modules.
 
+### MAJOR-004: Network Security Modules ✅
+- **File**: `src/security/network_security.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Refactored 780-line monolithic file into modular policies, monitoring, threats, and configuration modules. Orchestrator now coordinates components while tests pass.
+
 ### MODERATE-001: AI Development Workflow Refactor ✅
 - **File**: `src/ai_ml/dev_workflow.py`
 - **Status**: Completed
