+++ src/security/network_security_config.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python3
+"""Configuration constants for network security modules."""
+
+# Default settings for network scanning
+DEFAULT_MAX_THREADS = 100
+DEFAULT_TIMEOUT = 5
+
+# Anomaly detection threshold
+DEFAULT_ANOMALY_THRESHOLD = 0.8
+
+# Threat intelligence update interval (seconds)
+THREAT_UPDATE_INTERVAL = 3600
+
+__all__ = [
+    "DEFAULT_MAX_THREADS",
+    "DEFAULT_TIMEOUT",
+    "DEFAULT_ANOMALY_THRESHOLD",
+    "THREAT_UPDATE_INTERVAL",
+]
