+++ src/security/network_security_threats.py
@@ -1,0 +1,355 @@
+#!/usr/bin/env python3
+"""Threat assessment and intelligence modules."""
+
+import logging
+import time
+from typing import Dict, List, Optional
+
+from .network_security_policies import SecurityEvent
+from .network_security_config import THREAT_UPDATE_INTERVAL
+
+
+class VulnerabilityAssessor:
+    """Vulnerability assessment and CVE lookup system."""
+
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+        self.vulnerability_db: Dict = {}
+        self.cve_cache: Dict[str, Dict] = {}
+
+    def load_vulnerability_database(self) -> Dict:
+        """Load vulnerability database from file or API."""
+        try:
+            db_path = "vulnerabilities.db"
+            if hasattr(self, "_load_from_file"):
+                self.vulnerability_db = self._load_from_file(db_path)
+            else:
+                self.vulnerability_db = {
+                    "common_ports": {
+                        21: {"service": "FTP", "vulnerabilities": ["anonymous_access", "weak_auth"]},
+                        22: {"service": "SSH", "vulnerabilities": ["weak_keys", "default_creds"]},
+                        23: {"service": "Telnet", "vulnerabilities": ["cleartext_auth", "no_encryption"]},
+                        80: {"service": "HTTP", "vulnerabilities": ["sql_injection", "xss", "csrf"]},
+                        443: {"service": "HTTPS", "vulnerabilities": ["weak_crypto", "expired_certs"]},
+                        3306: {"service": "MySQL", "vulnerabilities": ["weak_auth", "no_ssl"]},
+                        5432: {"service": "PostgreSQL", "vulnerabilities": ["weak_auth", "no_ssl"]},
+                    }
+                }
+            self.logger.info(
+                f"Loaded {len(self.vulnerability_db)} vulnerability records"
+            )
+            return self.vulnerability_db
+        except Exception as e:
+            self.logger.error(f"Failed to load vulnerability database: {e}")
+            return {}
+
+    def scan_service(self, target: str, port: int, service: str) -> List[Dict]:
+        """Scan specific service for vulnerabilities."""
+        vulnerabilities: List[Dict] = []
+        try:
+            if port in self.vulnerability_db.get("common_ports", {}):
+                port_info = self.vulnerability_db["common_ports"][port]
+                for vuln_type in port_info["vulnerabilities"]:
+                    vulnerability = {
+                        "type": vuln_type,
+                        "service": service,
+                        "port": port,
+                        "target": target,
+                        "severity": self._assess_vulnerability_severity(vuln_type),
+                        "description": self._get_vulnerability_description(vuln_type),
+                        "recommendation": self._get_vulnerability_recommendation(vuln_type),
+                    }
+                    vulnerabilities.append(vulnerability)
+            service_vulns = self._check_service_specific_vulnerabilities(target, port, service)
+            vulnerabilities.extend(service_vulns)
+            self.logger.info(
+                f"Found {len(vulnerabilities)} vulnerabilities for {service} on {target}:{port}"
+            )
+            return vulnerabilities
+        except Exception as e:
+            self.logger.error(f"Service vulnerability scan failed: {e}")
+            return []
+
+    def lookup_cve(self, cve_id: str) -> Dict:
+        """Look up CVE information from database or API."""
+        try:
+            if cve_id in self.cve_cache:
+                return self.cve_cache[cve_id]
+            cve_info = {
+                "id": cve_id,
+                "description": f"Security vulnerability in {cve_id}",
+                "severity": "high",
+                "cvss_score": 8.5,
+                "affected_versions": ["1.0.0", "1.1.0"],
+                "references": [f"https://cve.mitre.org/cgi-bin/cvename.cgi?name={cve_id}"],
+                "published_date": "2021-12-10",
+                "last_updated": "2021-12-15",
+            }
+            self.cve_cache[cve_id] = cve_info
+            self.logger.info(f"Retrieved CVE information for {cve_id}")
+            return cve_info
+        except Exception as e:
+            self.logger.error(f"CVE lookup failed for {cve_id}: {e}")
+            return {
+                "id": cve_id,
+                "description": "CVE lookup failed",
+                "severity": "unknown",
+                "error": str(e),
+            }
+
+    def _assess_vulnerability_severity(self, vuln_type: str) -> str:
+        high_severity = ["sql_injection", "rce", "privilege_escalation"]
+        medium_severity = ["weak_auth", "xss", "csrf"]
+        low_severity = ["information_disclosure", "weak_crypto"]
+        if vuln_type in high_severity:
+            return "high"
+        if vuln_type in medium_severity:
+            return "medium"
+        if vuln_type in low_severity:
+            return "low"
+        return "medium"
+
+    def _get_vulnerability_description(self, vuln_type: str) -> str:
+        descriptions = {
+            "sql_injection": "SQL injection vulnerability allows malicious SQL code execution",
+            "xss": "Cross-site scripting vulnerability allows malicious script execution",
+            "weak_auth": "Weak authentication mechanism that can be easily bypassed",
+            "no_encryption": "Service communication is not encrypted",
+            "anonymous_access": "Anonymous access is allowed without authentication",
+        }
+        return descriptions.get(vuln_type, f"Unknown vulnerability type: {vuln_type}")
+
+    def _get_vulnerability_recommendation(self, vuln_type: str) -> str:
+        recommendations = {
+            "sql_injection": "Use parameterized queries and input validation",
+            "xss": "Implement proper output encoding and CSP headers",
+            "weak_auth": "Implement strong authentication and MFA",
+            "no_encryption": "Enable TLS/SSL encryption for all communications",
+            "anonymous_access": "Disable anonymous access and require authentication",
+        }
+        return recommendations.get(vuln_type, "Review and implement security best practices")
+
+    def _check_service_specific_vulnerabilities(self, target: str, port: int, service: str) -> List[Dict]:
+        vulnerabilities: List[Dict] = []
+        try:
+            if service.lower() in ["http", "https", "nginx", "apache"]:
+                vulnerabilities.extend(self._check_http_vulnerabilities(target, port))
+            elif service.lower() in ["mysql", "postgresql", "mongodb"]:
+                vulnerabilities.extend(
+                    self._check_database_vulnerabilities(target, port, service)
+                )
+            elif service.lower() == "ssh":
+                vulnerabilities.extend(self._check_ssh_vulnerabilities(target, port))
+        except Exception as e:
+            self.logger.debug(f"Service-specific vulnerability check failed: {e}")
+        return vulnerabilities
+
+    def _check_http_vulnerabilities(self, target: str, port: int) -> List[Dict]:
+        vulnerabilities: List[Dict] = []
+        try:
+            import requests
+            protocol = "https" if port == 443 else "http"
+            url = f"{protocol}://{target}:{port}"
+            response = requests.get(url, timeout=5, verify=False)
+            headers = response.headers
+            if "X-Frame-Options" not in headers:
+                vulnerabilities.append(
+                    {
+                        "type": "missing_security_header",
+                        "header": "X-Frame-Options",
+                        "severity": "medium",
+                        "description": "Missing clickjacking protection header",
+                        "recommendation": "Add X-Frame-Options header",
+                    }
+                )
+            if "X-Content-Type-Options" not in headers:
+                vulnerabilities.append(
+                    {
+                        "type": "missing_security_header",
+                        "header": "X-Content-Type-Options",
+                        "severity": "low",
+                        "description": "Missing MIME type sniffing protection",
+                        "recommendation": "Add X-Content-Type-Options: nosniff header",
+                    }
+                )
+        except Exception as e:
+            self.logger.debug(f"HTTP vulnerability check failed: {e}")
+        return vulnerabilities
+
+    def _check_database_vulnerabilities(self, target: str, port: int, service: str) -> List[Dict]:
+        return [
+            {
+                "type": "database_exposure",
+                "service": service,
+                "severity": "high",
+                "description": f"Database service {service} is exposed to network",
+                "recommendation": "Restrict database access to internal network only",
+            }
+        ]
+
+    def _check_ssh_vulnerabilities(self, target: str, port: int) -> List[Dict]:
+        return [
+            {
+                "type": "ssh_exposure",
+                "service": "SSH",
+                "severity": "medium",
+                "description": "SSH service is exposed to network",
+                "recommendation": "Use key-based authentication and restrict access",
+            }
+        ]
+
+
+class ThreatIntelligence:
+    """Threat intelligence and feed management system."""
+
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+        self.threat_feeds: List[Dict] = []
+        self.threat_cache: Dict[str, Dict] = {}
+        self.last_update = 0
+        self.update_interval = THREAT_UPDATE_INTERVAL
+
+    def get_latest_threats(self) -> List[Dict]:
+        """Get latest threats from intelligence feeds."""
+        try:
+            current_time = time.time()
+            if current_time - self.last_update > self.update_interval:
+                self._update_threat_feeds()
+                self.last_update = current_time
+            return list(self.threat_cache.values())
+        except Exception as e:
+            self.logger.error(f"Failed to get latest threats: {e}")
+            return []
+
+    def _update_threat_feeds(self) -> None:
+        """Update threat intelligence feeds (mock implementation)."""
+        mock_threat = {
+            "id": f"threat_{int(time.time())}",
+            "type": "malware",
+            "severity": "medium",
+            "description": "Mock threat feed entry",
+            "source": "local_intelligence",
+        }
+        self.threat_cache[mock_threat["id"]] = mock_threat
+
+    def check_ip_reputation(self, ip_address: str) -> Dict:
+        """Check IP reputation using intelligence feeds."""
+        try:
+            if ip_address in self.threat_cache:
+                return self.threat_cache[ip_address]
+            reputation = {
+                "ip_address": ip_address,
+                "reputation_score": 0.5,
+                "threat_level": "unknown",
+                "source": "local_intelligence",
+            }
+            self.threat_cache[ip_address] = reputation
+            return reputation
+        except Exception as e:
+            self.logger.error(f"IP reputation check failed for {ip_address}: {e}")
+            return {
+                "ip_address": ip_address,
+                "reputation_score": 0.5,
+                "threat_level": "unknown",
+                "error": str(e),
+            }
+
+
+class IncidentResponse:
+    """Incident response and handling system."""
+
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+        self.incident_history: List[Dict] = []
+        self.response_procedures: Dict = {}
+        self.escalation_levels = ["low", "medium", "high", "critical"]
+
+    def handle_event(self, event: SecurityEvent) -> bool:
+        """Handle security event and trigger appropriate response."""
+        try:
+            incident = {
+                "id": f"incident_{int(time.time())}",
+                "event": event,
+                "timestamp": time.time(),
+                "status": "open",
+                "response_actions": [],
+                "escalation_level": self._determine_escalation_level(event.severity),
+            }
+            self.incident_history.append(incident)
+            response_triggered = self._trigger_automated_response(incident)
+            self.logger.warning(
+                f"Security incident created: {incident['id']} - {event.event_type}"
+            )
+            return response_triggered
+        except Exception as e:
+            self.logger.error(f"Incident handling failed: {e}")
+            return False
+
+    def _determine_escalation_level(self, severity: str) -> str:
+        severity_mapping = {
+            "low": "low",
+            "medium": "medium",
+            "high": "high",
+            "critical": "critical",
+        }
+        return severity_mapping.get(severity, "medium")
+
+    def _trigger_automated_response(self, incident: Dict) -> bool:
+        try:
+            event = incident["event"]
+            response_actions: List[Dict] = []
+            if event.severity in ["high", "critical"]:
+                block_action = self._block_ip_address(event.source_ip)
+                if block_action:
+                    response_actions.append(block_action)
+            alert_action = self._generate_security_alert(incident)
+            if alert_action:
+                response_actions.append(alert_action)
+            incident["response_actions"] = response_actions
+            self.logger.info(
+                f"Automated response triggered for incident {incident['id']}"
+            )
+            return True
+        except Exception as e:
+            self.logger.error(f"Automated response failed: {e}")
+            return False
+
+    def _block_ip_address(self, ip_address: str) -> Dict:
+        try:
+            block_action = {
+                "type": "ip_block",
+                "target": ip_address,
+                "action": "block",
+                "timestamp": time.time(),
+                "status": "completed",
+            }
+            self.logger.info(f"IP address {ip_address} blocked")
+            return block_action
+        except Exception as e:
+            self.logger.error(f"IP blocking failed for {ip_address}: {e}")
+            return {}
+
+    def _generate_security_alert(self, incident: Dict) -> Dict:
+        try:
+            alert_action = {
+                "type": "security_alert",
+                "incident_id": incident["id"],
+                "severity": incident["escalation_level"],
+                "message": f"Security incident {incident['id']} requires attention",
+                "timestamp": time.time(),
+                "status": "sent",
+            }
+            self.logger.info(
+                f"Security alert generated for incident {incident['id']}"
+            )
+            return alert_action
+        except Exception as e:
+            self.logger.error(f"Alert generation failed: {e}")
+            return {}
+
+
+__all__ = [
+    "VulnerabilityAssessor",
+    "ThreatIntelligence",
+    "IncidentResponse",
+]
